Error parsing and syntax in the code were improved

+++ week5/diary/src/com/controllers/AddressController.java
@@ -28,24 +28,23 @@ public class AddressController extends Controllers {
     
     @Override
     public void create(String[] content) {
-        /* --------- Content array contains ---------
-            content[0] -> address name
-            content[1] -> foreign key (city id)
-        */
         Exceptions.checkAddressData(content);
+        String addressValue = content[0];
+        int city_id = Integer.parseInt(content[1]);
+        
         String sql = "INSERT INTO Address (name,city_id) VALUES (?,?);";
         int address_id = -1;
         try {
             PreparedStatement prstmt;
             prstmt = DBManagement.getConnection().prepareStatement(sql);
-            prstmt.setString(1,content[0]);
-            prstmt.setInt(2,Integer.parseInt(content[1]));
+            prstmt.setString(1,addressValue);
+            prstmt.setInt(2,city_id);
             prstmt.executeUpdate();
             prstmt.close();
-            System.out.println("Register was created successfully");
+            System.out.println("Address register was created successfully");
         }
         catch(SQLException e) {
-            System.err.println("Was not possible create register");
+            System.err.println("Was not possible create address register");
             return;
         }
         
@@ -67,13 +66,18 @@ public class AddressController extends Controllers {
 
         Address address = new Address();
         address.setAddress_id(address_id);
-        address.setName(content[0]);
-        address.setCity_id(Integer.parseInt(content[1]));
+        address.setName(addressValue);
+        address.setCity_id(city_id);
         objects.put(Integer.toString(address_id), address);
     }
     
     @Override
     public void delete(int id) {
+        if(!objects.containsKey(id)) {
+            System.err.println("Address map does not have the specified key!");
+            return;
+        }
+        
         String sql = "DELETE FROM Address WHERE address_id=" + id + ";";
         try {
             ResultSet rs = DBManagement.getStatement().executeQuery(sql);
@@ -85,24 +89,23 @@ public class AddressController extends Controllers {
             return;
         }
         
-        if(!objects.containsKey(id)) {
-            System.err.println("Address map does not have the specified key!");
-            return;
-        }
-        
         objects.remove(id);
     }
     
     @Override
     public void edit(int id, String[] content) {
-        /* --------- Content array contains ---------
-            content[0] -> address name
-            content[1] -> foreign key (city id)
-        */
+        if(!objects.containsKey(id)) {
+            System.err.println("Address map does not have the specified key!");
+            return;
+        }
+        
         Exceptions.checkAddressData(content);
+        String addressValue = content[0];
+        int city_id = Integer.parseInt(content[1]);
+        
         String sql = "UPDATE Address SET "
-            + "name='" + content[0] + "',"
-            + "city_id='" + content[1] + "'"
+            + "name='" + addressValue + "',"
+            + "city_id='" + city_id + "'"
             + " WHERE address_id=" + id + ";";
         try {
             ResultSet rs = DBManagement.getStatement().executeQuery(sql);
@@ -114,15 +117,10 @@ public class AddressController extends Controllers {
             return;
         }
         
-        if(!objects.containsKey(id)) {
-            System.err.println("Address map does not have the specified key!");
-            return;
-        }
-        
         Address address = (Address) objects.get(id);
         address.setAddress_id(id);
-        address.setName(content[0]);
-        address.setCity_id(Integer.parseInt(content[1]));
+        address.setName(addressValue);
+        address.setCity_id(city_id);
     }
     
     @Override
